services/etl/pyproject.toml
[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[project]
name = "helix-etl"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "pandas>=2.0.0",
    "numpy>=1.26.0",
    "ta>=0.11.0",
    "psycopg2-binary>=2.9.9",
    "structlog>=24.1.0",
]

[tool.setuptools.packages.find]
where = ["src"]
